moved database dump processing flow
